requirements.txt
joblib==0.14.1
numpy==1.18.1
opencv-python==4.1.2.30
pandas==0.25.3
Pillow==6.1.0
requests==2.22.0
scikit-learn==0.19.2
scipy==1.4.1
torch==1.3.1+cpu
torchvision==0.4.2+cpu
tqdm==4.32.2
